+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 0.7.1 (12-5-24)
 ---
 - set 1.18/ai-gateway and /portal-only envs to use `1.18.0-rc3` release
+- reconfigure ollama to own namespace in 1.18/ai-gateway/lb-failover-demo
+- add demo-script.sh in 1.18/ai-gateway/lb-failover-demo
 
 0.7.0 (11-19-24)
 ---